\name{ThatsOurIbex-package}
\alias{ThatsOurIbex-package}
\alias{ThatsOurIbex}
\docType{package}
\title{
\packageTitle{ThatsOurIbex}
}
\description{
Generates a random letter of the alphabet; it does not actually generate ibexes. Ibices?
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ThatsOurIbex}
\packageIndices{ThatsOurIbex}
Library function, generates random alphabet letters
}
\author{
\packageAuthor{ThatsOurIbex}

Maintainer: \packageMaintainer{ThatsOurIbex}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
library()
}
